Minor fixes to make the project build more easily
Apparently newer versions of g++ will attempt to build PIEs
(position-independent code) automatically unless you tell them not to,
which is a problem because DCLXVI is not position-independent. I added
-no-pie to CFLAGS to fix this, and also fixed a few sources of compiler
and Make warnings.

+++ lib/suffix/SuffixTree.cpp
@@ -327,7 +327,7 @@ void SuffixTree::validate()
         }
     if ( error == 0 )
         cout << "All Suffixes present!\n";
-    int leaf_count = 0;
+    size_t leaf_count = 0;
     int branch_count = 0;
     for ( auto& countEntry : branchCount ) {
         if ( countEntry.second == 0 )
